4.7.3

+++ src/BiliLite.UWP/Modules/Player/PlayerVM.cs
@@ -295,7 +295,7 @@ namespace BiliLite.Modules
     {
         public int node_id { get; set; }
         public string title { get; set; }
-        public int cid { get; set; }
+        public long cid { get; set; }
     }
 
     public class InteractionModel
